requirements.txt
Django
django-environ
Pillow
python-decouple
pytest-django
pytest-factoryboy
